requirements.txt
Flask==2.3.2
Werkzeug>=2.3.3
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
icrawler
jinja2